fix(m5): address Copilot review on PR #30
- Move PreprocessingConfig to config/preprocessing_config.py (new file)
  so the config layer has no dependency on augment/preprocessing.py and
  scipy/soundfile are not imported at config-schema load time
- augment/preprocessing.py now imports PreprocessingConfig from config/
  and re-exports it via __all__ so existing callers are unaffected
- preprocess() docstring: document the new config parameter
- cli.py: fix inline comment "peak-normalizes" → "peak-limits" to match
  the _peak_limit() limiter behaviour (never scales up)
- docs/audio_generation_v3_design.md: fix M5 row — "dataclass" → "Pydantic
  BaseModel"; list actual wiener_denoise flag
