+++ docs/audio_generation_v3_design.md
@@ -36,7 +36,7 @@ All P0–P2 work is tracked here. Milestones that span multiple PRs are listed o
 | **M3a** | P0 | ✅ Done | Preserves within-scene loudness trajectory | `rms_target_dbfs` on `StyleEntry`; `_apply_rms_gain()` helper in `SceneMixer`; 4-tuple segment API; AGG −28→−15 dBFS (I1→I5), VIC −26→−30 dBFS; FLOAT temp WAV write to avoid PCM_16 hard-clip | Small | No |
 | **M3b** | P0 | ✅ Done | Prevents peak-normalize from erasing inter-scene RMS contrast | Replace `_normalize_peak()` in `preprocess()` with sample-peak limiter (only clips above −1.0 dBFS, no forced scale-up); update `validate_audio()` to check peak ≤ −1.0 dBFS instead of == −1.0 dBFS; update `spec.md §3` | Small | Yes — spec.md §3 |
 | **M4** | P0 | ✅ Done | Fixes silent label corruption in strong-label JSONL | Audit `debug_run_1` emotional states; extend `taxonomy.yaml` `emotional_states` to cover all legitimate LLM outputs; replace `_normalize_emotion()` silent fallback with explicit mapping table + logged warning (hard fail for completely unknown states); add clip-level `quality_flags` value `"emotion_downgrade"` | Small | No |
-| **M5** | P1 | ✅ Done | Protects Tier A phonetic quality from over-denoising | `PreprocessingConfig` dataclass; update `preprocess()` to accept it (default unchanged); scene YAMLs gain optional `preprocessing` block; Wiener step skipped when `wiener_denoise=False`; unit tests verify Wiener skip | Small | No |
+| **M5** | P1 | ✅ Done | Protects Tier A phonetic quality from over-denoising | `PreprocessingConfig` Pydantic `BaseModel` with `wiener_denoise` flag; update `preprocess()` to accept it (default preserves current Wiener behavior); scene YAMLs gain optional `preprocessing` block; Wiener step skipped when `wiener_denoise=False`; unit tests verify Wiener skip | Small | No |
 | **M6** | P1 | 🔲 Not started | Replaces mechanical silence gaps with psychologically-motivated turn latencies | `TurnGapController` in `synthbanshee/tts/gap_controller.py`; project-specific gap tables (§4.5); wire into `TTSRenderer.render_scene()` replacing `turn.pause_before_s`; NEG/NEU confusor gap ranges; unit tests per context type and project | Small | No |
 | **M7** | P1 | 🔲 Not started | Eliminates speaker-state reset at turn boundaries | `SpeakerState` with `update()` / `to_metadata_dict()`; intensity-drift rules per role; wire into `TTSRenderer.render_scene()` — one `SpeakerState` per speaker; state offsets applied to render params; state written to per-turn generation metadata | Medium | No |
 | **M2b** | P1 | 🔲 Not started | Burst-pause microstructure within high-intensity turns | `PhraseHint` / `PhraseProsody` dataclasses in `ssml_types.py`; `DialogueTurn.phrase_hints`; char-offset resolver (maps `text_original` offsets → `text_spoken` after niqqud); `SSMLBuilder` extension for nested `<prosody>` + `<break>`; Jinja2 template update requesting `phrase_hints` for I3–I5 turns; deterministic sentence-final imperative heuristics | Medium | No |
