+++ synthbanshee/augment/preprocessing.py
@@ -18,8 +18,9 @@ from pathlib import Path
 
 import numpy as np
 import soundfile as sf
-from pydantic import BaseModel
 from scipy.signal import butter, resample_poly, sosfilt, wiener
+
+from synthbanshee.config.preprocessing_config import PreprocessingConfig
 
 _TARGET_SR = 16_000
 _TARGET_CHANNELS = 1
@@ -30,18 +31,8 @@ _SILENCE_PAD_S = 0.5  # minimum silence padding (seconds)
 _MIN_DURATION_S = 3.0  # clips below this are invalid (spec §3)
 _MAX_DURATION_S = 300.0  # clips above this must be segmented (spec §3)
 
-
-class PreprocessingConfig(BaseModel):
-    """Per-scene preprocessing options.
-
-    Passed to ``preprocess()`` to control which pipeline steps are applied.
-    All flags default to the spec-mandated behaviour so existing call sites
-    that omit the config are unaffected.
-    """
-
-    wiener_denoise: bool = True
-    """Apply Wiener noise reduction (step 4).  Set to False for Tier A scenes
-    where the clean TTS signal should be preserved without spectral smoothing."""
+# Re-export so callers can do: from synthbanshee.augment.preprocessing import PreprocessingConfig
+__all__ = ["PreprocessingConfig", "PreprocessingResult", "preprocess", "validate_audio"]
 
 
 @dataclass
@@ -129,6 +120,10 @@ def preprocess(
         output_path: Destination for the processed 16 kHz mono WAV.
         dirty_dir: If given, copy the raw input here before processing
                    (to retain the "dirty" original as required by spec §3.1).
+        config: Optional per-scene preprocessing flags.  ``None`` uses
+                ``PreprocessingConfig()`` defaults (all steps enabled).
+                Currently controls ``wiener_denoise`` — set to ``False``
+                to skip the Wiener filter for clean TTS Tier A scenes.
 
     Returns:
         PreprocessingResult with metadata about the processing applied.
