+++ synthbanshee/cli.py
@@ -282,8 +282,8 @@ def _run_generate_pipeline(
         with tempfile.TemporaryDirectory() as tmp:
             raw_wav = Path(tmp) / "raw.wav"
             # Write as 32-bit float WAV so out-of-range samples from M3 RMS
-            # gain are preserved exactly; preprocess() peak-normalizes and
-            # then writes the final PCM_16 output safely.
+            # gain are preserved exactly; preprocess() peak-limits any
+            # over-range peaks and then writes the final PCM_16 output safely.
             sf.write(str(raw_wav), mixed.samples, mixed.sample_rate, subtype="FLOAT")
             result = preprocess(raw_wav, clip_wav, dirty_dir=dirty_dir, config=scene.preprocessing)
     except Exception as exc:
